maps to symbols
This function will reconstruct a list of symbols that refer to a list
of keymaps

Good for reversing results from `current-active-maps' into a list of
useful symbols.

Had to strip sparse keymaps and switch to eq comparison to get good
results.  Looks like no false positives survived, as we would like.

Signed-off-by: Psionik K <73710933+psionic-k@users.noreply.github.com>