Make report rendering support recursive sections
Basically, if plist's :rows element contains more :rows, indent and
render the rows.

Together with section headers, this gives us a basic form of rendering
multiple levels of results hierarchies, which is good for things like
using keymap names as section headers for a list of bindings as occurs
when scanning for stupid keys.

Signed-off-by: Psionik K <73710933+psionic-k@users.noreply.github.com>