implement ignored maps
Some maps are just really funky and trying to make downstream survive the weird
inputs is a pain.

Other maps, such as function key and key translation, are somewhat harmless in
that they only map keys to keys.  They could be offensive to some users, but it
is supported via customize

Signed-off-by: Psionik K <73710933+psionic-k@users.noreply.github.com>