treat esc-map like an extension of global map
The global map an esc-map have some strange relationship that requires special
casing them.  Lookups for meta keys in the global map are fine.  Scans however
will not see any sequences that are only present in the esc-map.  This change
presents these sequences during the global map scan, as the meta-prefixed
sequences.

Signed-off-by: Psionik K <73710933+psionic-k@users.noreply.github.com>