! introducing matching on starts-with-key-or-modified
First off, what was done...

This commit has to temporarily complicate some things to get settled.  To
support matching multiple entries per map, the shadows was made a bit more
generic.  The pattern for later refactor is beginning to emerge.

The shadows commands were all made to call render directly, removing an annoying
aspect of the interface, which is still not that well settled.

...and where it's going

All reporting functions are basically using maps, predicates, and some slight
variations in processing the result of keymap scans or producing a row using a
lookup.

This is also relatively close to the declarations necessary for a mass-unbinding
DSL.  Expressions of keymaps, predicates, exceptions, and actions are basically
a mass-unbinding language.

I think the pattern for refactor has emerged to create some generic definitions
for everything.  The interface won't be exactly like keymap-QL or magit, but it
will be pretty freaking generic and customizeable / extensible by going down
this road.

Signed-off-by: Psionik K <73710933+psionic-k@users.noreply.github.com>